refactor: single user only - remove client list
