+++ package.json
@@ -20,8 +20,6 @@
   },
   "devDependencies": {
     "@castlenine/svelte-qrcode": "^2.3.0",
-    "@dicebear/collection": "^9.2.2",
-    "@dicebear/core": "^9.2.2",
     "@eslint/js": "^9.19.0",
     "@playwright/experimental-ct-svelte": "1.55.0",
     "@playwright/test": "1.55.0",
