+++ src/lib/components/pdf-export.svelte
@@ -17,16 +17,16 @@
   interface Props {
     graphValueData: GraphPortfolioValueType
     adjustWithInflation: boolean
-    clientBirthDate?: Date
+    birthDate?: Date
     portfolioName: string
-    clientName?: string
+    profileName?: string
     keyDates?: Date[]
   }
 
   let {
     graphValueData,
     adjustWithInflation,
-    clientBirthDate,
+    birthDate,
     portfolioName,
     keyDates = [],
   }: Props = $props()
@@ -151,7 +151,7 @@
         </div>
 
         <div class="full-chart-container">
-          <GraphPortfolioValue {graphValueData} {adjustWithInflation} {clientBirthDate} />
+          <GraphPortfolioValue {graphValueData} {adjustWithInflation} {birthDate} />
         </div>
 
         <PdfExportFooter pageNumber={1} {totalPages} />
@@ -183,7 +183,7 @@
                         investments={graphValueData?.investments || []}
                         currency={graphValueData?.portfolio?.currency || 'USD'}
                         currentYear={breakdown.year}
-                        {clientBirthDate}
+                        {birthDate}
                         {adjustWithInflation}
                       />
                     </div>
