+++ src/lib/routes.ts
@@ -6,30 +6,23 @@ const routePrefix = `${base}${ROUTER === 'hash' ? '/#' : ''}`
 export default {
   HOME: `${routePrefix}/`,
   SETTINGS: `${routePrefix}/settings`,
-  NEW_CLIENT: `${routePrefix}/client/new`,
-  CLIENT: (id: string) => `${routePrefix}/client/${id}`,
-  CLIENT_NEW_PORTFOLIO: (clientId: string) => `${routePrefix}/client/${clientId}/portfolio/new`,
-  CLIENT_PORTFOLIO: (clientId: string, portfolioId: string) =>
-    `${routePrefix}/client/${clientId}/portfolio/${portfolioId}`,
-  CLIENT_EDIT_PORTFOLIO: (clientId: string, portfolioId: string) =>
-    `${routePrefix}/client/${clientId}/portfolio/${portfolioId}/edit`,
-  EDIT_CLIENT: (id: string) => `${routePrefix}/client/${id}/edit`,
-  NEW_INVESTMENT: (clientId: string, portfolioId: string) =>
-    `${routePrefix}/client/${clientId}/portfolio/${portfolioId}/new-investment`,
-  EDIT_INVESTMENT: (clientId: string, portfolioId: string, investmentId: string) =>
-    `${routePrefix}/client/${clientId}/portfolio/${portfolioId}/edit-investment/${investmentId}`,
-  NEW_GOAL: (clientId: string, portfolioId: string) =>
-    `${routePrefix}/client/${clientId}/portfolio/${portfolioId}/goals/new`,
-  RETIREMENT_GOAL_CALCULATOR: (clientId: string, portfolioId: string) =>
-    `${routePrefix}/client/${clientId}/portfolio/${portfolioId}/goals/retirement/calculator`,
-  RETIREMENT_GOAL_PREVIEW: (clientId: string, portfolioId: string) =>
-    `${routePrefix}/client/${clientId}/portfolio/${portfolioId}/goals/retirement/preview`,
-  KID_EDUCATION_GOAL_CALCULATOR: (clientId: string, portfolioId: string) =>
-    `${routePrefix}/client/${clientId}/portfolio/${portfolioId}/goals/kid-education/calculator`,
-  KID_EDUCATION_GOAL_PREVIEW: (clientId: string, portfolioId: string) =>
-    `${routePrefix}/client/${clientId}/portfolio/${portfolioId}/goals/kid-education/preview`,
-  PDF_EXPORT_PARAMS: (clientId: string, portfolioId: string) =>
-    `${routePrefix}/client/${clientId}/portfolio/${portfolioId}/pdf-export-params`,
-  PDF_EXPORT: (clientId: string, portfolioId: string) =>
-    `${routePrefix}/client/${clientId}/portfolio/${portfolioId}/pdf-export`,
+  EDIT_PROFILE: `${routePrefix}/profile/edit`,
+  NEW_PORTFOLIO: `${routePrefix}/portfolio/new`,
+  PORTFOLIO: (portfolioId: string) => `${routePrefix}/portfolio/${portfolioId}`,
+  EDIT_PORTFOLIO: (portfolioId: string) => `${routePrefix}/portfolio/${portfolioId}/edit`,
+  NEW_INVESTMENT: (portfolioId: string) => `${routePrefix}/portfolio/${portfolioId}/new-investment`,
+  EDIT_INVESTMENT: (portfolioId: string, investmentId: string) =>
+    `${routePrefix}/portfolio/${portfolioId}/edit-investment/${investmentId}`,
+  NEW_GOAL: (portfolioId: string) => `${routePrefix}/portfolio/${portfolioId}/goals/new`,
+  RETIREMENT_GOAL_CALCULATOR: (portfolioId: string) =>
+    `${routePrefix}/portfolio/${portfolioId}/goals/retirement/calculator`,
+  RETIREMENT_GOAL_PREVIEW: (portfolioId: string) =>
+    `${routePrefix}/portfolio/${portfolioId}/goals/retirement/preview`,
+  KID_EDUCATION_GOAL_CALCULATOR: (portfolioId: string) =>
+    `${routePrefix}/portfolio/${portfolioId}/goals/kid-education/calculator`,
+  KID_EDUCATION_GOAL_PREVIEW: (portfolioId: string) =>
+    `${routePrefix}/portfolio/${portfolioId}/goals/kid-education/preview`,
+  PDF_EXPORT_PARAMS: (portfolioId: string) =>
+    `${routePrefix}/portfolio/${portfolioId}/pdf-export-params`,
+  PDF_EXPORT: (portfolioId: string) => `${routePrefix}/portfolio/${portfolioId}/pdf-export`,
 }
