+++ src/lib/stores/app.svelte.ts
@@ -1,10 +1,17 @@
 import { SvelteSet } from 'svelte/reactivity'
-import type { ClientNested, ClientNoId, ClientStore } from '$lib/types'
-import { clientNestedSchema, storedDataSchema, type StoredData } from '$lib/schemas'
-import { withClientStore } from './client.svelte'
+import type { Portfolio, PortfolioNested, Profile } from '$lib/types'
+import type { PortfolioStore } from './portfolio.svelte'
+import { withPortfolioStore } from './portfolio.svelte'
+import { storedDataSchema, type StoredData } from '$lib/schemas'
 import { storageErrorStore } from './storage-error.svelte'
 
 const STORAGE_KEY = 'kalkul-data'
+
+const DEFAULT_PROFILE: Profile = {
+  name: '',
+  email: '',
+  birth_date: '',
+}
 
 function loadData(): StoredData {
   try {
@@ -15,18 +22,23 @@ function loadData(): StoredData {
   } catch (e) {
     console.error('Failed to load data from localStorage', e)
   }
-  return { lastUpdated: 0, clients: [] }
+  return { lastUpdated: 0, profile: { ...DEFAULT_PROFILE }, portfolios: [] }
 }
 
 function withAppStore() {
-  let clients = $state.raw<ClientStore[]>([])
+  let profile = $state<Profile>({ ...DEFAULT_PROFILE })
+  let portfolios = $state<PortfolioStore[]>([])
   let loading = $state(true)
   let lastUpdated = $state(0)
   const hiddenInvestmentIds = new SvelteSet<string>()
 
   function persist(): void {
     const now = Date.now()
-    const stored: StoredData = { lastUpdated: now, clients }
+    const stored: StoredData = {
+      lastUpdated: now,
+      profile,
+      portfolios: portfolios.map((p) => p.toJSON()),
+    }
     try {
       localStorage.setItem(STORAGE_KEY, JSON.stringify(stored))
       lastUpdated = now
@@ -35,37 +47,46 @@ function withAppStore() {
       console.error('Failed to save data to localStorage', e)
       storageErrorStore.setError()
     }
-    // Trigger reactivity: $state.raw only signals on reassignment, so create
-    // a new array reference after every mutation.
-    clients = [...clients]
+    // Trigger reactivity: $state reassignment
+    portfolios = [...portfolios]
   }
 
-  function deleteClient(id: string): void {
-    const idx = clients.findIndex((c) => c.id === id)
-    if (idx === -1) return
-    clients.splice(idx, 1)
+  function deletePortfolio(id: string): void {
+    const idx = portfolios.findIndex((p) => p.id === id)
+    if (idx !== -1) portfolios.splice(idx, 1)
     persist()
+  }
+
+  function duplicatePortfolio(newPortfolio: PortfolioNested): string | undefined {
+    const enrichedPortf = withPortfolioStore(newPortfolio, appParent)
+    portfolios.push(enrichedPortf)
+    persist()
+    return newPortfolio.id
   }
 
   const appParent = {
     persist,
-    deleteClient,
+    deletePortfolio,
+    duplicatePortfolio,
     hiddenIds: hiddenInvestmentIds,
   }
 
-  function enrichAll(rawClients: ClientNested[]): ClientStore[] {
-    return rawClients.map((c) => withClientStore(c, appParent))
+  function enrichAll(rawPortfolios: PortfolioNested[]): PortfolioStore[] {
+    return rawPortfolios.map((p) => withPortfolioStore(p, appParent))
   }
 
   return {
     get lastUpdated() {
       return lastUpdated
     },
-    get clients() {
-      return clients
+    get profile() {
+      return profile
     },
-    set clients(value: ClientStore[]) {
-      clients = value
+    get portfolios() {
+      return portfolios
+    },
+    set portfolios(value: PortfolioStore[]) {
+      portfolios = value
       loading = false
     },
     get loading() {
@@ -75,21 +96,46 @@ function withAppStore() {
       loading = value
     },
     reset() {
-      clients = []
+      profile = { ...DEFAULT_PROFILE }
+      portfolios = []
       loading = true
     },
 
     persist,
-    deleteClient,
+    deletePortfolio,
     get hiddenIds() {
       return hiddenInvestmentIds
+    },
+
+    // --- Profile ---
+
+    updateProfile(updates: Partial<Profile>) {
+      profile = { ...profile, ...updates }
+      persist()
+    },
+
+    // --- Portfolios ---
+
+    addPortfolio(data: Omit<Portfolio, 'id'>): string {
+      const portId = crypto.randomUUID()
+      const newPortfolio: PortfolioNested = {
+        ...data,
+        id: portId,
+        investments: [],
+        goals: [],
+      }
+      const enrichedPortf = withPortfolioStore(newPortfolio, appParent)
+      portfolios.push(enrichedPortf)
+      persist()
+      return portId
     },
 
     // --- Load ---
 
     load(): void {
       const data = loadData()
-      clients = enrichAll(data.clients)
+      profile = data.profile
+      portfolios = enrichAll(data.portfolios)
       lastUpdated = data.lastUpdated
       loading = false
     },
@@ -102,7 +148,8 @@ function withAppStore() {
           const data = storedDataSchema.parse(JSON.parse(event.newValue))
           if (data.lastUpdated === lastUpdated) return
 
-          clients = enrichAll(data.clients)
+          profile = data.profile
+          portfolios = enrichAll(data.portfolios)
           lastUpdated = data.lastUpdated
         } catch {
           // Ignore malformed data from other tabs
@@ -113,37 +160,17 @@ function withAppStore() {
       return () => window.removeEventListener('storage', onStorage)
     },
 
-    // --- Finders ---
-
-    findClient(id: string): ClientStore | undefined {
-      return clients.find((c) => c.id === id)
-    },
-
-    // --- Clients ---
-
-    addClient(client: ClientNoId): string {
-      const id = crypto.randomUUID()
-      const newClient: ClientNested = {
-        ...client,
-        id,
-        portfolios: [],
-      }
-      const enrichedClient = withClientStore(newClient, appParent)
-      clients.push(enrichedClient)
-      persist()
-      return id
-    },
-
     // --- Backup / Restore ---
 
     exportBackup(): string {
-      return JSON.stringify(clients, undefined, 2)
+      return JSON.stringify({ profile, portfolios }, undefined, 2)
     },
 
     importBackup(json: string): void {
       const parsed: unknown = JSON.parse(json)
-      const validated = clientNestedSchema.array().parse(parsed)
-      clients = enrichAll(validated)
+      const validated = storedDataSchema.pick({ profile: true, portfolios: true }).parse(parsed)
+      profile = validated.profile
+      portfolios = enrichAll(validated.portfolios)
       loading = false
       persist()
     },
