--- src/routes/(app)/client/[id]/portfolio/new/+page@(app).svelte
@@ -1,23 +1,0 @@
-<script lang="ts">
-  import { _ } from 'svelte-i18n'
-  import { page } from '$app/state'
-  import EditPortfolio from '$lib/components/edit-portfolio.svelte'
-  import Fullscreen from '$lib/components/fullscreen.svelte'
-  import Typography from '$lib/components/ui/typography.svelte'
-  import { appStore } from '$lib/stores/app.svelte'
-
-  const clientId = page.params.id
-  const client = $derived(appStore.findClient(clientId))
-
-  function close() {
-    history.back()
-  }
-</script>
-
-<Fullscreen>
-  {#if client}
-    <EditPortfolio {close} {client} />
-  {:else}
-    <Typography variant="h2">404 - {$_('common.notFound')}</Typography>
-  {/if}
-</Fullscreen>
