+++ src/routes/(app)/portfolio/new/+page@(app).svelte
@@ -1,0 +1,12 @@
+<script lang="ts">
+  import EditPortfolio from '$lib/components/edit-portfolio.svelte'
+  import Fullscreen from '$lib/components/fullscreen.svelte'
+
+  function close() {
+    history.back()
+  }
+</script>
+
+<Fullscreen>
+  <EditPortfolio {close} />
+</Fullscreen>
