+++ src/routes/(app)/profile/edit/+page@(app).svelte
@@ -1,0 +1,12 @@
+<script lang="ts">
+  import EditProfile from '$lib/components/edit-profile.svelte'
+  import Fullscreen from '$lib/components/fullscreen.svelte'
+
+  function close() {
+    history.back()
+  }
+</script>
+
+<Fullscreen>
+  <EditProfile {close} />
+</Fullscreen>
